Move fragments to subdirectories.
