+++ Android/app/src/main/res/layout/fragment_club_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.prototype1.view.ClubDetailFragment">
+    tools:context="com.example.prototype1.view.sideFragments.ClubDetailFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
