+++ Android/app/src/main/res/layout/fragment_event_list.xml
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#E0E0E0"
-        tools:context="com.example.prototype1.view.EventListFragment">
+        tools:context="com.example.prototype1.view.mainFragments.EventListFragment">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
